fix(trace): show prompt badge on observations again
